hcp/requirements.txt
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
boto3==1.34.100
azure-storage-blob==12.19.1
pyspark==3.5.1
pyarrow==16.0.0
openpyxl==3.1.2
requests==2.31.0
pyyaml==6.0.1
scikit-learn==1.4.2
jupyter==1.0.0
matplotlib==3.8.4
seaborn==0.13.2
python-dotenv==1.0.1
pytest==8.2.0
